moved plotting related stuff to own namespace
