+++ cellphonedb/src/api_endpoints/terminal_api/method_terminal_api_endpoints/method_terminal_commands.py
@@ -1,4 +1,3 @@
-import os
 import sys
 import traceback
 from typing import Optional, Any, Callable
@@ -20,6 +19,7 @@ from cellphonedb.src.exceptions.ParseCountsException import ParseCountsException
 from cellphonedb.src.exceptions.ParseMetaException import ParseMetaException
 from cellphonedb.src.exceptions.ReadFileException import ReadFileException
 from cellphonedb.src.local_launchers.local_method_launcher import LocalMethodLauncher
+from cellphonedb.src.plotters import r_plotter
 
 
 def check_subsampling_params(ctx: Context, argument: click.Argument, value) -> Any:
@@ -231,44 +231,4 @@ def analysis(meta_filename: str,
 @click.option('--columns', type=click.File('r'), help='List of columns to plot, one per line [all available]')
 @click.option('--plot-function', type=str, default='dot_plot', help='Plot function name in R code [dot_plot]')
 def plot(results_path, rows, columns, plot_function):
-    os.chdir(results_path)
-
-    means_df = pd.read_csv('./means.txt', sep='\t')
-    n_rows, n_cols = means_df.shape
-    n_cols -= 9
-
-    n_rows, selected_rows = selected_items(rows, n_rows)
-    n_cols, selected_columns = selected_items(columns, n_cols)
-
-    this_file_dir = os.path.dirname(os.path.realpath(__file__))
-
-    robjects.r.source(os.path.join(this_file_dir, 'plotters/plot_dot_by_column_name.R'))
-    available_names = list(robjects.globalenv.keys())
-
-    if plot_function in available_names:
-        function_name = plot_function
-    else:
-        raise MissingPlotterFunctionException()
-
-    robjects.r('library(ggplot2)')
-
-    plotter = robjects.r[function_name]
-
-    plotter(width=int(5 + max(3, n_cols * 0.8)),
-            height=int(5 + max(5, n_rows * 0.5)),
-            selected_rows=selected_rows,
-            selected_columns=selected_columns
-            )
-
-
-def selected_items(selection: Optional[click.File], size):
-    if selection is not None:
-        df = pd.read_csv(selection, header=None)
-        names = df[0].tolist()
-
-        selected = StrVector(names)
-        size = len(names)
-    else:
-        selected = robjects.NULL
-
-    return size, selected
+    r_plotter.plot(plot_function, results_path, rows, columns)
