+++ cellphonedb/src/plotters/R/plot_dot_by_column_name.R
